rest_log: move exceptions to their own file
